% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1_fusion.R
\name{.genFusedLassoProximal}
\alias{.genFusedLassoProximal}
\title{Fused lasso optimisation with the proximal-gradient method (Chen et al. 2010).
Internal method}
\usage{
.genFusedLassoProximal(X, Y, groups, group.names = sort(unique(groups)),
  XX = NULL, XY, p, k, samp.sizes, lambda, gamma, C, G, epsilon = 1e-04
  * p * dim(C)[2], tol = 1e-06, num.it = 1000, lam.max = NULL,
  intercept = TRUE, penalty.factors = NULL, X.list = NULL,
  c.flag = FALSE)
}
\arguments{
\item{X}{matrix of covariates (n by p)}

\item{Y}{vector of responses (length n)}

\item{groups}{vector of group indicators (length n)}

\item{group.names}{group names (length k)}

\item{XX}{\code{t(X) \%*\% X}}

\item{XY}{\code{t(X) \%*\% Y}}

\item{p}{Number of covariates}

\item{k}{Number of groups}

\item{samp.sizes}{Sample size for each group}

\item{lambda}{Sparsity hyperparameter}

\item{gamma}{Fusion hyperparameter}

\item{C}{constraints}

\item{G}{pairwise fusion hyperparameters (tau)}

\item{epsilon}{step size}

\item{tol}{tolerance}

\item{num.it}{number of iterations}

\item{lam.max}{maximal eigenvalue (pre-calculated)}

\item{intercept}{whether to include intercept or not}

\item{penalty.factors}{setting l1 penalty factors for different coefficients}

\item{X.list}{X as a list (required if lam.max calculated on the fly)}

\item{c.flag}{C++ subroutines will be used if true.}
}
\description{
Fused lasso optimisation with the proximal-gradient method (Chen et al. 2010).
Internal method
}
